+++ src/app/loading.tsx
@@ -1,18 +1,5 @@
-import ResumeCardLoad from '@/components/server/main/ResumeCardLoad'
+import ResumeCardListLoader from '@/components/ui/loading/ResumeCardLoad'
 
 export default function Loading() {
-  return (
-    <section className="grid w-full grid-cols-2 gap-6 max-sm:grid-cols-1">
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-      <ResumeCardLoad />
-    </section>
-  )
+  return <ResumeCardListLoader />
 }
