+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import ResumeCardList from '@/components/client/main/ResumeCardList'
+import ResumeCardList from '@/components/main/ResumeCardList'
 
 export default function Home() {
   return <ResumeCardList />
