+++ src/app/resume/edit/page.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-underscore-dangle */
 import AuthorizedAccess from '@/utils/AuthorizedAccess'
-import CreateProfile from '@/components/client/resumeld/CreateProfile'
+import CreateProfile from '@/components/editor/CreateProfile'
 import { getServerSession } from 'next-auth'
 import { authOptions } from '@/pages/api/auth/[...nextauth]'
 import { redirect } from 'next/navigation'
