+++ src/app/resume/page.tsx
@@ -4,7 +4,7 @@ import { authOptions } from '@/pages/api/auth/[...nextauth]'
 import { getServerSession } from 'next-auth'
 import getQueryClient from '@/utils/getQueryClient'
 import AuthorizedAccess from '@/utils/AuthorizedAccess'
-import ResumeDetails from '@/components/client/resume/ResumeDetail'
+import ResumeDetails from '@/components/resume/ResumeDetail'
 import { Hydrate, dehydrate } from '@tanstack/react-query'
 import { getPost } from '@/lib/api/post'
 import { redirect } from 'next/navigation'
