+++ src/components/Editor/UserProfileInputs.tsx
@@ -1,8 +1,8 @@
 import { useFieldArray, useFormContext, Controller } from 'react-hook-form'
 import { IResumeData } from '@/types/resumeDataType'
-import Input from '@/components/client/ui/Input'
-import InputError from '@/components/client/ui/InputError'
-import Txt from '@/components/client/ui/Txt'
+import Input from '@/components/ui/Input'
+import InputError from '@/components/ui/InputError'
+import Txt from '@/components/ui/Txt'
 import { FaMinus, FaPlus } from 'react-icons/fa'
 import { phoneReg, emailReg } from '@/lib/constants/regex'
 import React, { useState, useRef, ChangeEvent, Fragment } from 'react'
