+++ src/components/resume/ResumeDetail.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import Category from '@/components/server/resume/Category'
-import Profile from '@/components/server/resume/Profile'
-import Skill from '@/components/server/resume/Skill'
-import CategoryDetails from '@/components/server/resume/CategoryDetails'
+import Category from '@/components/resume/Category'
+import Profile from '@/components/resume/Profile'
+import Skill from '@/components/resume/Skill'
+import CategoryDetails from '@/components/resume/CategoryDetails'
 import { Session } from 'next-auth'
 import { useQuery } from '@tanstack/react-query'
 import { getPost } from '@/lib/api/post'
@@ -27,13 +27,13 @@ export default function ResumeDetails({ session, id }: IResumeDetailProps) {
         <Profile profileData={data.userInfo} />
         {/* 자기소개 */}
         <Category title="Introduce">
-          <Category.Direction direction="none">
+          <Category.Direction direction="none" padding="1">
             <p className="w-full gap-1">{data.userInfo.introduction}</p>
           </Category.Direction>
         </Category>
         {/* 스킬 셋 */}
         <Category title="Skill Set">
-          <Category.Direction direction="horizon">
+          <Category.Direction direction="horizontal" padding="2/4" gap="gap-4">
             {data.techStack?.map((el) => (
               <Skill key={el.title} title={el.title} description={el.description} />
             ))}
@@ -43,7 +43,7 @@ export default function ResumeDetails({ session, id }: IResumeDetailProps) {
         {data.categorys?.map((el) => {
           return (
             <Category title={el.title} key={el.title}>
-              <Category.Direction direction="vertical">
+              <Category.Direction direction="vertical" padding="2/4" gap="gap-4">
                 {el.detail.map((detail) => {
                   return <CategoryDetails detailData={detail} key={detail.title} />
                 })}
